feat: 레포지토리 추상화
